+++ lib/src/analyzers/unused_l10n_analyzer/reporters/unused_l10n_report_params.dart
@@ -1,0 +1,6 @@
+/// Represents additional unused l10n reporter params
+class UnusedL10NReportParams {
+  final bool congratulate;
+
+  const UnusedL10NReportParams({required this.congratulate});
+}