+++ lib/src/cli/cli_runner.dart
@@ -2,10 +2,12 @@ import 'dart:io';
 
 import 'package:args/command_runner.dart';
 
+import '../version.dart';
 import 'commands/analyze_command.dart';
 import 'commands/check_unused_code_command.dart';
 import 'commands/check_unused_files_command.dart';
 import 'commands/check_unused_l10n_command.dart';
+import 'models/flag_names.dart';
 
 /// Represents a cli runner responsible
 /// for running a command based on raw cli call data.
@@ -17,12 +19,15 @@ class CliRunner extends CommandRunner<void> {
     CheckUnusedCodeCommand(),
   ];
 
-  CliRunner()
-      : super(
-          'metrics',
-          'Analyze and improve your code quality.',
-        ) {
+  late final IOSink _output;
+
+  CliRunner([IOSink? output])
+      : super('metrics', 'Analyze and improve your code quality.') {
+    _output = output ?? stdout;
+
     _commands.forEach(addCommand);
+
+    _usesVersionOption();
   }
 
   /// Represents the invocation string message
@@ -32,14 +37,26 @@ class CliRunner extends CommandRunner<void> {
   /// Main entry point for running a command
   @override
   Future<void> run(Iterable<String> args) async {
+    final results = parse(args);
+    final showVersion = results[FlagNames.version] as bool;
+
+    if (showVersion) {
+      _output.writeln('Dart Code Metrics version: $packageVersion');
+
+      return;
+    }
+
     try {
       await super.run(_addDefaultCommand(args));
     } on UsageException catch (e) {
-      print('${e.message}\n');
-      print('${e.usage}\n');
+      _output
+        ..writeln(e.message)
+        ..writeln(e.usage);
+
       exit(64);
     } on Exception catch (e) {
-      print('Oops; metrics has exited unexpectedly: "$e"');
+      _output.writeln('Oops; metrics has exited unexpectedly: "$e"');
+
       exit(1);
     }
   }
@@ -49,4 +66,14 @@ class CliRunner extends CommandRunner<void> {
       : !commands.keys.contains(args.first)
           ? ['analyze', ...args]
           : args;
+
+  void _usesVersionOption() {
+    argParser
+      ..addSeparator('')
+      ..addFlag(
+        FlagNames.version,
+        help: 'Reports the version of this tool.',
+        negatable: false,
+      );
+  }
 }