+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## Unreleased
 
 * feat: add static code diagnostics `format-comment`.
-* chore: restrict `analyzer` version to `>=2.4.0 <3.3.0`.
+* feat: add new command flag `--no-congratulate`.
+* feat: add `--version` flag to print current version of the package.
 * fix: make `check-unused-l10n` also cover supertype member calls.
+* fix: cyclomatic complexity calculation for functions with internal lambdas.
+* chore: restrict `analyzer` version to `>=2.4.0 <3.3.0`.
 
 ## 4.10.0-dev.2